requirements.txt
qdrant-client>=1.15.0
python-dotenv>=1.0.0
mistralai>=1.0.0
elevenlabs>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0

